[refactor: monitor struct]
Merge the overall monitor area and area available to render windows and
give the respective members more descriptive names.

With the updated statusbar, the monitor area available for windows is no
longer reduced by the height of the statusbar. Hence, the overall
monitor area is equal to the area for rendering windows, thus both can
be merged.