[events: fix monitor discovery]
issue:

Changes to the monitor setup are not detected, hence the internal
monitor representation is not updated.

reason:

The ConfigureNotify event had been disabled with an earlier change.

fix:

Add the handler for the ConfigureNotify event.

Update the xinerama monitor discovery routine. Since clients are no
longer bound to a monitor, the removal of a monitor or more
specifically, the loop that was supposed to detach clients from the
monitor that shall be removed and attach them to a new monitor, resulted
in an infinite loop since clients are detached and attached to the
global dwm client list over and over.