[main: overhaul event loop]
Replace the main event loop, which just checked for xlib events, with an
epoll-based loop, allowing to react to more than just xlib events.

Currently, the updated loop is required for a timer which handles the
modifier key state reset. Unfortunately, the xlib KeyRelease event does
not report all events for released modifier keys, causing the actually
keyboard and the dwm state to be out of sync. As a workaround a timer is
used to track and reset the modifier key state.