[input: rm button remap on focus change]
Remove changing the button mappings for a client on focus change.

Thus far button mappings for clients have been update when the focus
changes, by removing all mappings, adding them again and depending on
the client's focus state, add an ANY-map. Removing and adding the user
mappings doesn't seem to be meaningful, since they do not change.
Removing them seems to be needed to get rid of the ANY-map. However, it
the reason for the ANY-map is not documented and everything works also
without it, except changing client focus using the mouse. This can,
however, be solved by using a nop mapping.