[actions: fix window maximise toggle]
issue:

When maximising the dimensions of a client one after the other, only the
second dimension can be toggled back to its original value.

reason:

client_resize() stores all values of the original client geometry before
each resize. Hence, the original value of the first changed dimension is
overwritten with the maximised value when the second dimension is also
maximised.

fix:

Only store those parts of the geometry that actually change.