[actions: fix client maximise]
issue:

1. Clients do not get maximised.
2. Clients cannot be toggled between maximised and original value.
3. Layouts do not use the entire monitor area.
4. Changing the size of a window in either dimension back and forth
   does not restore the initial geometry.

reason:

Before resizing a window, win_resize() applies the size hints, which
alters the resulting geometry.

For one, this causes the window to not actually being resized to the
dimensions of the respective monitor, causing issues 1 and 3.

Secondly, this disables the toggle features of action_client_resize(),
since it compares the client dimensions against the monitor dimensions,
only toggling a dimension if it is equal to the monitors dimensions.

fix:

Do not apply window size hints if the client resize operation is
triggered by a user action or layout.