[actions: fix client edge movement]
issue:

Moving clients to the edges of a secondary monitor and the edge snapping
do not work if the client is not on the primary monitor.

reason:

action_client_move() and action_client_move_mouse() always use the first
of the discovered monitors.

fix:

Update both functions to use the monitor a client is mapped to.