[clients: update init of transients]
Initialise transient windows in the same loop as non-transient ones,
since there is no apparent reason to separate the two. The only effect
of the two loops is the order in which the windows are initially put
onto the stack. Since the order is constantly changed, this doesn't seem
a meaningful reason to handle them differently.

With this change, also override_redirect clients are now ignored
entirely. Previously, the would have been managed if they were also
transient windows.