[dmenu: integrate into repo]
Add the dmenu build to the repo build system.

The dmenu code itself is not overhauled, even though the functions in
drw and util look pretty similar to the ones already present in dwm.

Update the dmenu path and run scripts to also work if they are installed
in a directory which is not part of PATH.

Extend dmenurun to use exec instead of a shell for executing the command
selected via dmenu. This avoids having an open shell process.