[core: SIGCHLD handling]
Register a proper signal handler for SIGCHLD to avoid zombie processes.

Thus far zombies could be avoided by ignoring SIGCHLD and setting the
correct signal handler flags. However, with this approach the exit codes
of child processes cannot be acquired. Hence, calls to e.g. system() do
not return the childs exit code but -1. Since the statdctrl refresh in
the statusbar relies on the statdctrl exit code to determine if statd is
running an actual signal handler is needed to both, avoid zombie
processes and be able to acquire child process exit codes.