[dmenu: fix statusbar overlay]
issue:

The statusbar gets raised over the dmenurun window.

reason:

The statusbar timer regularly checks if statd is running. If it is not
running, the timer handler resets the WM_NAME attribute, which is used
to set the statusbar status text, which in turn triggers a statusbar
update, that is raising the statusbar window.

fix:

Instead of simply forking dmenu, fork it and wait for it to finish. This
blocks dwm from any further actions as long as dmenu is still open.

However, since dmenurun, thus far, used exec to execute the selected
application, the now replaced dmenurun process does not return until the
application returns. To fix this
 - extend dmenu to fork-exec the selected item instead of printing it
 - update dmenurun to use the new dmenu option