[mappings: fix buttons not working in application]
issue:

Left mouse button not working in client windows, e.g. in vlc or a
browser the left mouse button doesn't do anything.

reason:

The mapping for the left button without any modifiers is registered to
be reported to dwm to allow changing the window focus using the mouse.
As an effect the event is not sent to the clients, hence they cannot
process it.

fix:

Add a modifier to the button mapping.

The original dwm updated the registered buttons each time the client
focus changes. With the refactoring this was changed to only register
buttons with a client on its initialisation.
To allow changing the window focus with a mouse button without a
modifier being pressed, it would be needed to reintroduce the original
dwm solution. However, to avoid those function calls on each focus
change, it has been decided to stick with the current implementation and
rather use a modifier for the focus change button mapping.