[commands: Add client maximisation]
Add the client maximise command, which is somewhat of a light wight
fullscreen mode. It removes the window borders and maximises it
horizontally and vertically, setting the related xlib properties.

Some applications, react independently and differently to the horizontal
and vertical maximisation than they react to the fullscreen property.
For example, both, opera and firefox, render videos on their entire
window area if the fullscreen xlib property is set. However, only
firefox removes its tab and address bars if the property is set.
On the other hand, opera removes its rounded corners if the horizontal
and vertical maximisation properties are set and doesn't show any
further reaction to the fullscreen property.

The new client command especially allows the user to control the
annoyingly rounded corners of opera -- removing them when wanting to
render a video over the entire screen.