[commands: add size hints to client move & resize]
Update the client (re)size and move commands to leverage the client's
size hints.

Thus far CONFIG_WIN_MOVE_DELTA is set such, that it is a multiple of the
number of pixels per text line of an xterm window to avoid having xterms
which cannot use the entire window height.

By using a clients size hints, this xterm-specific solution can be
avoided, since the hints already contain the information on valid size
increments as well as min and max values.