[commands: avoid layout arrangement calls]
Disable client move and resize commands when the active layout is not a
floating one, i.e. provides an arrange callback. Hence, calling
layout_arrany() at the end of those commands is not meaningful anymore.

Layouts are meant to control the size and position of clients, hence
they  shall not be changed by the user.