[core: add restart via signal]
Add a signal handler for SIGINT which triggers a restart of dwm.

For an yet unknown reason dwm sometimes stops to react to or receive
keyboard inputs. Thus far dwm had to be killed an restarted, killing all
open applications with it. To avoid killing the applications, SIGINT can
now be used to externally trigger a dwm restart.