[core: fix no keys]
issue:

From time to time, after closing an application, keyboard events are no
longer delivered to dwm. Depending on the configuration, this might
render dwm completely unresponsive, not able to open applications or
restart itself.

dwm itself is still running an responding to events, such as timers or
button events.

reason:

The reason is not totally clear, but seems to be an issue with the
client focus. In particular, if the last application on a tag is closed,
the focus remains on the last, now deleted, client.

It is unclear if keeping the focus on a non-existing client is really
the issue, since the issue cannot be reproduced reliably. For instance,
closing xterm as the last application on a tag doesn't cause the issue.
Most often it could be reproduced with jameica. However, also with
jameica the issue didn't appear all the time.

fix:

Ensure the root window is focused if no other client is available.