[core: add desktop atoms]
Add support for the NET_CURRENT_DESKTOP and NET_WM_DESKTOP atoms, which
are used to indicate the current desktop and the desktop a client is on.

Some tools like xdo use both attributes to query the client state. xdo
in particular is used in the systools shell scripts and reports an error
if the atoms are not supported by the window manager.