[core: fix window move requests]
issue:

If a client application moves a window twice to the same location, one
would assume that the window is only moved once, keeping its location
for the second move.
However, if the window is on a monitor other than the one with origin
(0, 0) the window is actually moved twice, whereas the second location
is not the specified one.

reason:

The ConfigureRequest handler performs move requests relative to a
window's current monitor. Hence, the window being moved from the monitor
with origin (0, 0) to another monitor works as intended. However, the
second move to that same location is then relative to the second
monitor, hence moving the window again.

fix:

Interpret the ConfigureRequest coordinates as relative to (0, 0)
instead of a window's monitor.