[clients: use (un)map to show/hide clients]
Use a client's mapping state to show/hide its window, instead of moving
the window outside of the visible screen space.

The old method of moving windows out of sight to hide them, doesn't work
well with applications that move windows themselves. Showing/hiding
windows in dwm is for instance used to show/hide the windows for the
selected tags. Hence, if an application would move a window, which is
not supposed to be visible for the current tag mask, to the visible
screen area, the window would be visible to the user even though it is
not supposed to be.

To avoid this conflict clients are unmapped to hide them. That way
applications can still move windows, without making them visible.

As a consequence, the MapRequest handler needed to be updated to no
longer remove a client if it gets unmapped.